Day 2 Functioning claw and picking up toys
The claw has movement and capable of grabbing toys.
Toys can be grabbed.

The claw and toy use temp assets

The claw can move with A/D or Left and Right arrow keys
The claw will descend with Space Bar
For testing purposes, Z will drop the toy the claw is holding